Move CAL-ADDRESS type to new data type system
