+++ ical/_types.py
@@ -24,7 +24,6 @@ ParsedProperty.
 from __future__ import annotations
 
 import copy
-import dataclasses
 import datetime
 import enum
 import json
@@ -32,17 +31,16 @@ import logging
 from collections.abc import Callable
 from typing import Any, Generator, Optional, TypeVar, Union, get_args, get_origin
 
-from pydantic import BaseModel, Field, root_validator
+from pydantic import BaseModel, root_validator
 from pydantic.dataclasses import dataclass
 from pydantic.fields import SHAPE_LIST, ModelField
 
 from .parsing.component import ParsedComponent
-from .parsing.property import ParsedProperty, ParsedPropertyParameter
+from .parsing.property import ParsedProperty
 from .recur import Recur
-from .types import Uri
 from .types.boolean import BooleanEncoder
 from .types.const import Classification, EventStatus, JournalStatus, TodoStatus
-from .types.data_types import DATA_TYPE, encode_model_property_params
+from .types.data_types import DATA_TYPE
 from .types.date_time import DateTimeEncoder
 from .types.duration import DurationEncoder
 from .types.integer import IntEncoder
@@ -104,72 +102,6 @@ def parse_parameter_values(cls: BaseModel, values: dict[str, Any]) -> dict[str, 
                     raise ValueError("Unexpected repeated property parameter")
                 values[param["name"]] = param["values"][0]
     return values
-
-
-class CalAddress(BaseModel):
-    """A value type for a property that contains a calendar user address.
-
-    This is a subclass of string so that it can be used in place of a string
-    to get the calendar address, but also supports additional properties.
-    """
-
-    uri: Uri = Field(alias="value")
-    """The calendar user address as a uri."""
-
-    common_name: Optional[str] = Field(alias="CN", default=None)
-    """The common name associated with the calendar user."""
-
-    user_type: Optional[str] = Field(alias="CUTYPE", default=None)
-    """The type of calendar user specified by the property.
-
-    Common values are defined in CalendarUserType, though also supports other
-    values not known by this library so it uses a string.
-    """
-
-    delegator: Optional[list[Uri]] = Field(alias="DELEGATED-FROM", default=None)
-    """The users that have delegated their participation to this user."""
-
-    delegate: Optional[list[Uri]] = Field(alias="DELEGATED-TO", default=None)
-    """The users to whom the user has delegated participation."""
-
-    directory_entry: Optional[Uri] = Field(alias="DIR", default=None)
-    """Reference to a directory entry associated with the calendar user."""
-
-    member: Optional[list[Uri]] = Field(alias="MEMBER", default=None)
-    """The group or list membership of the calendar user."""
-
-    status: Optional[str] = Field(alias="PARTSTAT", default=None)
-    """The participation status for the calendar user."""
-
-    role: Optional[str] = Field(alias="ROLE", default=None)
-    """The participation role for the calendar user."""
-
-    rsvp: Optional[bool] = Field(alias="RSVP", default=None)
-    """Whether there is an expectation of a favor of a reply from the calendar user."""
-
-    sent_by: Optional[Uri] = Field(alias="SENT-BY", default=None)
-    """Specifies the calendar user is acting on behalf of another user."""
-
-    language: Optional[str] = Field(alias="LANGUAGE", default=None)
-
-    _parse_parameter_values = root_validator(pre=True, allow_reuse=True)(
-        parse_parameter_values
-    )
-
-    @classmethod
-    def __encode_property_value__(cls, model_data: dict[str, str]) -> str | None:
-        return model_data.pop("value")
-
-    @classmethod
-    def __encode_property_params__(
-        cls, model_data: dict[str, Any]
-    ) -> list[ParsedPropertyParameter]:
-        return encode_model_property_params(cls.__fields__.values(), model_data)
-
-    class Config:
-        """Pyandtic model configuration."""
-
-        allow_population_by_field_name = True
 
 
 @dataclass
@@ -251,12 +183,6 @@ class PropertyDataType(enum.Enum):
     # Types to support
     #   BINARY
     #   TIME
-    CAL_ADDRESS = (
-        "CAL-ADDRESS",
-        CalAddress,
-        dataclasses.asdict,
-        None,  # Uses pydantic jason BaseModel encoder
-    )
     RECUR = (
         "RECUR",
         Recur,
@@ -404,7 +330,6 @@ class ComponentModel(BaseModel):
         _LOGGER.debug("Parsing value data %s", values)
 
         for field in cls.__fields__.values():
-            _LOGGER.debug("field=%s", field)
             if field.alias == "extras":
                 continue
             if not (value := values.get(field.alias)):
@@ -413,7 +338,6 @@ class ComponentModel(BaseModel):
                 # The incoming value is not from the parse tree
                 continue
             validators = _get_validators(field.type_)
-            _LOGGER.debug("validators=%s", validators)
             validated = []
             for prop in value:
                 # This property value may contain repeated values itself
