+++ ical/journal.py
@@ -10,10 +10,10 @@ from typing import Any, Optional, Union
 
 from pydantic import Field, root_validator
 
-from ._types import CalAddress, ComponentModel, RequestStatus, validate_until_dtstart
+from ._types import ComponentModel, RequestStatus, validate_until_dtstart
 from .parsing.property import ParsedProperty
 from .recur import Recur
-from .types import Uri
+from .types import CalAddress, Uri
 from .types.const import Classification, JournalStatus
 from .util import dtstamp_factory, normalize_datetime, uid_factory
