+++ ical/todo.py
@@ -7,17 +7,11 @@ from typing import Any, Optional, Union
 
 from pydantic import Field, root_validator
 
-from ._types import (
-    CalAddress,
-    ComponentModel,
-    Priority,
-    RequestStatus,
-    validate_until_dtstart,
-)
+from ._types import ComponentModel, Priority, RequestStatus, validate_until_dtstart
 from .alarm import Alarm
 from .parsing.property import ParsedProperty
 from .recur import Recur
-from .types import Geo, Uri
+from .types import CalAddress, Geo, Uri
 from .types.const import Classification, TodoStatus
 from .util import dtstamp_factory, normalize_datetime, uid_factory
