+++ ical/types/cal_address.py
@@ -1,0 +1,88 @@
+"""Library for parsing and encoding CAL-ADDRESS values."""
+
+from __future__ import annotations
+
+import dataclasses
+import logging
+from typing import Any, Optional
+
+from pydantic import BaseModel, Field, root_validator
+
+from ical.parsing.property import ParsedProperty, ParsedPropertyParameter
+
+from .data_types import DATA_TYPE, encode_model_property_params
+from .parsing import parse_parameter_values
+from .uri import Uri
+
+_LOGGER = logging.getLogger(__name__)
+
+
+@DATA_TYPE.register("CAL-ADDRESS")
+class CalAddress(BaseModel):
+    """A value type for a property that contains a calendar user address.
+    This is a subclass of string so that it can be used in place of a string
+    to get the calendar address, but also supports additional properties.
+    """
+
+    uri: Uri = Field(alias="value")
+    """The calendar user address as a uri."""
+
+    common_name: Optional[str] = Field(alias="CN", default=None)
+    """The common name associated with the calendar user."""
+
+    user_type: Optional[str] = Field(alias="CUTYPE", default=None)
+    """The type of calendar user specified by the property.
+    Common values are defined in CalendarUserType, though also supports other
+    values not known by this library so it uses a string.
+    """
+
+    delegator: Optional[list[Uri]] = Field(alias="DELEGATED-FROM", default=None)
+    """The users that have delegated their participation to this user."""
+
+    delegate: Optional[list[Uri]] = Field(alias="DELEGATED-TO", default=None)
+    """The users to whom the user has delegated participation."""
+
+    directory_entry: Optional[Uri] = Field(alias="DIR", default=None)
+    """Reference to a directory entry associated with the calendar user."""
+
+    member: Optional[list[Uri]] = Field(alias="MEMBER", default=None)
+    """The group or list membership of the calendar user."""
+
+    status: Optional[str] = Field(alias="PARTSTAT", default=None)
+    """The participation status for the calendar user."""
+
+    role: Optional[str] = Field(alias="ROLE", default=None)
+    """The participation role for the calendar user."""
+
+    rsvp: Optional[bool] = Field(alias="RSVP", default=None)
+    """Whether there is an expectation of a favor of a reply from the calendar user."""
+
+    sent_by: Optional[Uri] = Field(alias="SENT-BY", default=None)
+    """Specifies the calendar user is acting on behalf of another user."""
+
+    language: Optional[str] = Field(alias="LANGUAGE", default=None)
+
+    _parse_parameter_values = root_validator(pre=True, allow_reuse=True)(
+        parse_parameter_values
+    )
+
+    @classmethod
+    def __parse_property_value__(cls, prop: ParsedProperty) -> dict[str, str]:
+        """Convert the property into a dictionary for pydantic model."""
+        # XXX???
+        return dataclasses.asdict(prop)
+
+    @classmethod
+    def __encode_property_value__(cls, model_data: dict[str, str]) -> str | None:
+        return model_data.pop("value")
+
+    @classmethod
+    def __encode_property_params__(
+        cls, model_data: dict[str, Any]
+    ) -> list[ParsedPropertyParameter]:
+        return encode_model_property_params(cls.__fields__.values(), model_data)
+
+    class Config:
+        """Pyandtic model configuration."""
+
+        allow_population_by_field_name = True
